move axelar utility to axelar
